GPIO-Proximity_module-master/requirements.txt
Package    Version
---------- -------
paho-mqtt  1.6.1
pip        22.3.1
RPi.GPIO   0.7.1
setuptools 65.6.3
wheel      0.38.4
